+++ assets/javascripts/kitten/components/carousel/nav-tab-carousel/nav-tab-carousel.stories.js
