+++ assets/javascripts/kitten/components/carousel/simple-carousel/index.js
@@ -1,10 +1,10 @@
 import React, { Component, createRef } from 'react'
 import PropTypes from 'prop-types'
 import styled, { css } from 'styled-components'
-import COLORS from '../../constants/colors-config'
-import { createRangeFromZeroTo } from '../../helpers/utils/range'
-import { pxToRem } from '../../helpers/utils/typography'
-import { VisuallyHidden } from '../../components/accessibility/visually-hidden'
+import COLORS from '../../../constants/colors-config'
+import { createRangeFromZeroTo } from '../../../helpers/utils/range'
+import { pxToRem } from '../../../helpers/utils/typography'
+import { VisuallyHidden } from '../../../components/accessibility/visually-hidden'
 
 const StyledContainer = styled.div`
   ${({ addBottomMargin }) =>
